new allocrunner identity_hook
